% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{appendPlaceholder}
\alias{appendPlaceholder}
\title{Append placeholder cell types onto the reference}
\usage{
appendPlaceholder(reference, numPlaceholder = 1)
}
\arguments{
\item{reference}{original reference matrix}

\item{numPlaceholder}{number of placeholder cell types wanted (Default: 1)}
}
\value{
New reference matrix with dimension (cell type + placeholder) x gene
}
\description{
Given current cell type x gene reference matrix, append placeholder
cell types with number specified by the user, default to be 1. Note that placeholder
will be generated using dirichlet distribution with uniform mean
}
\examples{

data(mouseHypothalamus)
new_reference <- appendPlaceholder(mouseHypothalamus$reference, numPlaceholder=1)

}
