% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{commonGeneFilter}
\alias{commonGeneFilter}
\title{For provided spatial data and scRNA-seq data, find intersecting gene set}
\usage{
commonGeneFilter(spatial, scRNAseq)
}
\arguments{
\item{spatial}{matrix representing spatial data with dimension pixel x gene}

\item{scRNAseq}{matrix representing scRNAseq data with dimension cell x gene}
}
\value{
spatial and scRNA-seq data after selecting only common gene set
}
\description{
For provided spatial data and scRNA-seq data, find intersecting gene set
}
\examples{

data(scRNAseq_to_reference)
common_geneset <- commonGeneFilter(scRNAseq_to_reference$spatial, scRNAseq_to_reference$scRNAseq)

}
