% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{corMatrixPlot}
\alias{corMatrixPlot}
\title{Gene expression profiles correlation plot}
\usage{
corMatrixPlot(
  deconGexp1,
  deconGexp2,
  title = "Transcriptional correlation",
  type = "pearson",
  rowlab = "Deconvolved cell-types",
  collab = "Ground truth cell-types"
)
}
\arguments{
\item{deconGexp1}{first cell type gene expression profile with dimension cell type x gene}

\item{deconGexp2}{second cell type gene expression profile with dimension cell type x gene}

\item{title}{title for the correlation plot (Default: "Transcriptional correlation")}

\item{type}{type of correlation plotted (Default: "pearson")}

\item{rowlab}{row label (Default: "Deconvolved cell-types")}

\item{collab}{column title (Default: "Ground truth cell-types")}
}
\value{
A correlation matrix heatmap plot
}
\description{
Plot correlation between two sets of gene expression profiles as
heatmap.
}
