% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{formReference}
\alias{formReference}
\title{Form reference based on processed scRNAseq data}
\usage{
formReference(scRNAseq, celltype, minCell = 25, emptyFill = 0.1)
}
\arguments{
\item{scRNAseq}{matrix representing scRNAseq data with dimension cell x gene}

\item{celltype}{named vector representing cell type annotation for every cell}

\item{minCell}{minimum number of cells required to keep current cell type}

\item{emptyFill}{we replace genes that are not expressed at all in a cell type
by a pseudocount (default: 0.1)}
}
\value{
reference with dimension cell type x gene
}
\description{
For given scRNAseq data and the cell type annotation for every
cell, form cell type x gene refernece.
}
\examples{

data(scRNAseq_to_reference)
common_geneset <- commonGeneFilter(scRNAseq_to_reference$spatial, scRNAseq_to_reference$scRNAseq)
redundant_removal <- redundantGeneFilter(common_geneset$spatial, common_geneset$scRNAseq)

}
