% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{getLegendGrob}
\alias{getLegendGrob}
\title{Create legend object for plotting deconvolution result}
\usage{
getLegendGrob(legend_labels, legend_colors)
}
\arguments{
\item{legend_labels}{Labels for the cell types}

\item{legend_colors}{Colors for the cell types}
}
\value{
A legend object if it exists
}
\description{
Create a pseudo ggplot2 plot using the provided label and
color, then retrieving only the legend object and return
}
\examples{

data(mouse_hypothalamus)
color = c('Astrocyte'='red', 'Endothelial'='orange',
          'Ependymal'='black', 'Excitatory'='blue',
          'Inhibitory'='green', 'Microglia'='purple',
          'OD Immature'='yellow', 'OD Mature'='brown',
          'Pericytes'='cyan')
legend_labels <- names(color)
legend_colors <- as.vector(color)
legend_grob <- getLegendGrob(legend_labels, legend_colors)
plotDeconvRes(mouse_hypothalamus$ground_truth_deconv,
              mouse_hypothalamus$spatial_meta,
              color = color,
              numRow = 1,
              numCol = 2,
              legend_grob = legend_grob)

}
