% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpretation.R
\name{identityAssignmentOther}
\alias{identityAssignmentOther}
\title{Assign non-original cell type identity to estimated cell types}
\usage{
identityAssignmentOther(
  estGeneExp,
  provGeneExp,
  deconvRes,
  unknownIdentity,
  assignedDeconvRes,
  assignedGeneExp,
  corMethod = "pearson",
  threshold = 0.8,
  verbose = T
)
}
\arguments{
\item{estGeneExp}{unscaled estimated cell type gene expression profile}

\item{provGeneExp}{provided cell type gene expression profile}

\item{deconvRes}{unscaled deconvolution result}

\item{unknownIdentity}{cell types with unknown identities so far}

\item{assignedDeconvRes}{unscaled estimated deconvolution result of cell types
with identity already assigned}

\item{assignedGeneExp}{unscaled estimated gene expression profile of cell types
with identity already assigned}

\item{corMethod}{correlation method (Default: "pearson")}

\item{threshold}{correlation threshold for assigning cell type identity (Default: 0.8)}

\item{verbose}{Default: T}
}
\value{
A list containing deconvolution result and gene expression profile for
cell types with identity assigned, as well as cell types with unknown identity
}
\description{
Assign identity to cell types whose estimated gene expression
profile is similar (in terms of correlation) to provided gene expression profile
of any other cell types
}
