% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{imperfectReferenceResult}
\alias{imperfectReferenceResult}
\title{Sample data used in the inference vignette}
\format{
It contains a deconvolution result from applying imperfect reference
to the mouse hypothalamus data, in the standard FlexiDeconv output format
}
\source{
\url{https://datadryad.org/dataset/doi:10.5061/dryad.8t8s248}
}
\usage{
data(imperfectReferenceResult)
}
\description{
Sample data used in the inference vignette
}
\keyword{datasets}
