% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mouseHypothalamus}
\alias{mouseHypothalamus}
\title{Sample data used in the FlexiDeconv vignette}
\format{
It contains a list having:
\enumerate{
\item Spatial Transcriptomics data generated from single-cell profiling of the
hypothalamic preoptic region of mouse.
\item Metadata for the Spatial Transcriptomics data (e.g., location).
\item Relevant reference matrix.
\item Ground truth gene expression profile for every cell type.
\item Ground truth cell type proportion for every pixel.
}
}
\source{
\url{https://datadryad.org/dataset/doi:10.5061/dryad.8t8s248}
}
\usage{
data(mouseHypothalamus)
}
\description{
Sample data used in the FlexiDeconv vignette
}
\keyword{datasets}
