% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{outlierCellRemoval}
\alias{outlierCellRemoval}
\title{Outlier cell removal in scRNAseq data}
\usage{
outlierCellRemoval(
  scRNAseq,
  celltypes,
  lowerThresholdRank = 2,
  lowerThresholdCount = 3,
  plot = T,
  plot_density = 5
)
}
\arguments{
\item{scRNAseq}{matrix representing scRNAseq data with dimension cell x gene}
}
\value{
scRNAseq data after removing outlier cells
}
\description{
For provided scRNAseq data, remove outlier cells. Outlier is
defined as:
\enumerate{
\item cells with total gene count that is 1.5 IQR above Q3 of all
cells within the current cell type.
\item cells with rank x (default: 2) having at least y (default: 5) count.
}
}
\examples{

data(scRNAseq_to_reference)
common_geneset <- commonGeneFilter(scRNAseq_to_reference$spatial, scRNAseq_to_reference$scRNAseq)

}
