% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotDeconvRes}
\alias{plotDeconvRes}
\title{Visualizing deconvolution output for Spatial Transcriptomics data that have
multiple slices}
\usage{
plotDeconvRes(
  deconvResult,
  meta,
  color,
  numRow,
  numCol,
  r = 40,
  title = "Cell Type Proportion per Pixel",
  legend_grob = NA,
  verbose = T,
  topic = NA
)
}
\arguments{
\item{deconvResult}{deconvolution result, dimension: pixel x cell type}

\item{meta}{metadata for the Spatial Transcriptomics data, row names must be
barcodes, it also should have x, y as two columns, as well as a column for slice}

\item{color}{named vector representing the color you want for each cell type}

\item{numRow}{number of row of grid}

\item{numCol}{number of column of grid}

\item{r}{radius for each pixel as a circle (Default: 40)}

\item{title}{title of the graph (Default: "Cell Type Proportion per Pixel")}

\item{legend_grob}{object that will be displayed as legend}

\item{verbose}{Default: T}

\item{topic}{set a number if you want to visualize a specific cell type (Default: NA)}
}
\value{
Piechart visualization for deconvolution result
}
\description{
Given the deconvolution output for Spatial Transcriptomics, plot
the result to a grid, every entry represents a single slice
}
\examples{

data(mouseHypothalamus)
color = c('Astrocyte'='red', 'Endothelial'='orange',
'Ependymal'='black', 'Excitatory'='blue',
'Inhibitory'='green', 'Microglia'='purple',
'OD Immature'='yellow', 'OD Mature'='brown', 'Pericytes'='cyan')
legend_grob <- getLegendGrob(color)
plotDeconvRes(mouseHypothalamus$ground_truth_deconv,
mouseHypothalamus$spatial_meta,
color = color,
numRow = 1,
numCol = 2,
legend_grob = legend_grob)

}
