% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{redundantGeneFilter}
\alias{redundantGeneFilter}
\title{Remove redundant genes}
\usage{
redundantGeneFilter(spatial, scRNAseq)
}
\arguments{
\item{spatial}{matrix representing spatial data with dimension pixel x gene}

\item{scRNAseq}{matrix representing scRNAseq data with dimension cell x gene}
}
\value{
spatial and scRNA-seq data after removing redundant genes
}
\description{
For given spatial data and scRNAseq data, remove genes that have
0 counts across all pixels and all cells.
}
\examples{

data(scRNAseq_to_reference)
common_geneset <- commonGeneFilter(scRNAseq_to_reference$spatial, scRNAseq_to_reference$scRNAseq)
redundant_removal <- redundantGeneFilter(common_geneset$spatial, common_geneset$scRNAseq)

}
