% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reference_ready}
\alias{reference_ready}
\title{Example for allowing user to input their own reference matrix}
\format{
It contains reference in the format of cell type x gene. The purpose
is to use this dataset to demonstrate that user can input their own reference.
}
\source{
\url{https://assets.nemoarchive.org/dat-61kfys3}
}
\usage{
data(reference_ready)
}
\description{
Example for allowing user to input their own reference matrix
}
\keyword{datasets}
