% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{replaceZero}
\alias{replaceZero}
\title{Replace 0s in the reference}
\usage{
replaceZero(reference, zeroFillFactor = 0.1)
}
\arguments{
\item{reference}{user-provided reference with dimension cell type x gene}

\item{zeroFillFactor}{user-configurable factor for determining the factor to
multiply the minimum non-zero value by, to replace 0s}
}
\value{
reference with dimension cell type x gene
}
\description{
For user-provided reference, replace 0 by 0.1 (default) times the
minimum non-zero value in the reference
}
\examples{

data(reference_ready)
ref_replace_zero <- replaceZero(reference_ready, zeroFillFactor = 0.1)

}
