% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlexiDeconv_algo.R
\name{runFlexiDeconv}
\alias{runFlexiDeconv}
\title{Run FlexiDeconv algorithm for cell type deconvolution}
\usage{
runFlexiDeconv(
  spatial,
  reference,
  prior_const,
  verbose = T,
  iter_print_freq = 50
)
}
\arguments{
\item{spatial}{pixel x gene matrix representing Spatial Transcriptomics data}

\item{reference}{cell type x gene matrix representing reference, row sum should be 1}

\item{prior_const}{vector of length = total number of provided cell type in the reference,
representing weight assigned to each cell type.}

\item{verbose}{(default: TRUE)}

\item{iter_print_freq}{(default: 50)}
}
\value{
A list containing estimated cell type proportion per pixel and estimated
gene expression profile for each cell type
}
\description{
Given the Spatial Transcriptomics data (pixel x gene) and
reference (cell type x gene), as well as prior constant for the reference, run
FlexiDeconv algorithm to obtain:
\enumerate{
\item Cell type proportion per pixel
\item Estimated gene expression profile for each cell type
}
}
\examples{

data(mouse_hypothalamus)
new_reference <- appendPlaceholder(mouse_hypothalamus$reference, numPlaceholder=1)
prior_const = c(rep(5, 8), 0.1)
output = runFlexiDeconv(as.matrix(mouse_hypothalamus$spatial),
new_reference,
prior_const)

}
