\name{scRNAseq_to_reference}
\alias{scRNAseq_to_reference}
\docType{data}
\title{
Sample data used in the scRNAseq to reference conversion procedurem
}
\description{
Sample data used in the scRNAseq to reference conversion procedurem
}
\usage{data(scRNAseq_to_reference)}
\format{
It contains scRNAseq data as well as the spatial data used to mimic
the procedure of generating reference based on scRNAseq data.
}

\source{
\url{https://assets.nemoarchive.org/dat-61kfys3}
}
\examples{
data(scRNAseq_to_reference)
}
\keyword{datasets}



