% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpretation.R
\name{unknownCluster}
\alias{unknownCluster}
\title{For the final unknown cell types, perform clustering}
\usage{
unknownCluster(
  estGeneExp,
  deconvRes,
  unknownIdentity,
  assignedDeconvRes,
  assignedGeneExp,
  corMethod = "pearson",
  threshold = 0.7,
  verbose = T
)
}
\arguments{
\item{estGeneExp}{unscaled estimated cell type gene expression profile}

\item{deconvRes}{unscaled deconvolution result}

\item{unknownIdentity}{cell types with unknown identities so far}

\item{assignedDeconvRes}{unscaled estimated deconvolution result of cell types
with identity already assigned}

\item{assignedGeneExp}{unscaled estimated gene expression profile of cell types
with identity already assigned}

\item{corMethod}{correlation method (Default: "pearson")}

\item{threshold}{correlation threshold for assigning cell type identity (Default: 0.7)}

\item{verbose}{Default: T}
}
\value{
A list containing deconvolution result and gene expression profile for
cell types with identity assigned, as well as cell types with unknown identity
}
\description{
For the final unknown cell types, perform clustering and report
final interpreted deconvolution result and gene expression profile
}
