% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{visualizeReference}
\alias{visualizeReference}
\title{Visualize gene expression profile of reference}
\usage{
visualizeReference(
  ref,
  want_log = F,
  cluster_col = T,
  cluster_row = F,
  title = "Gene Expression Heatmap"
)
}
\arguments{
\item{ref}{reference matrix}

\item{want_log}{whether visualization should be done in log scale (Default: F)}

\item{cluster_col}{whether genes should be reordered (Default: T)}

\item{cluster_row}{whether cell types should be reordered (Default: F)}

\item{title}{Default: Gene Expression Heatmap}
}
\value{
Heatmap visualisation of gene expression profiles of reference
}
\description{
Given current cell type x gene reference matrix, visualize it
using heatmap. Allow visualization to be done in log scale. In default,
it automatically reorders genes for better interpretation.
}
\examples{

data(mouse_hypothalamus)
new_reference <- appendPlaceholder(mouse_hypothalamus$reference, numPlaceholder=1)
p <- visualizeReference(new_reference)
p

}
