% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{getLegendGrob}
\alias{getLegendGrob}
\title{Create legend object for plotting deconvolution result}
\usage{
getLegendGrob(color)
}
\arguments{
\item{color}{named vector representing color for each cell type}
}
\value{
A legend object if it exists
}
\description{
Create a pseudo ggplot2 plot using the provided label and
color, then retrieving only the legend object and return
}
\examples{

data(mouseHypothalamus)
color = c('Astrocyte'='red', 'Endothelial'='orange',
          'Ependymal'='black', 'Excitatory'='blue',
          'Inhibitory'='green', 'Microglia'='purple',
          'OD Immature'='yellow', 'OD Mature'='brown',
          'Pericytes'='cyan')
legend_grob <- getLegendGrob(color)
plotDeconvRes(mouseHypothalamus$ground_truth_deconv,
              mouseHypothalamus$spatial_meta,
              color = color,
              numRow = 1,
              numCol = 2,
              legend_grob = legend_grob)

}
